Redid parts of EthNameModule, use NameManagerInterface as base
